Logo null

+++ src/main/java/utn/t2/s1/gestionsocios/controllers/DepartamentoController.java
@@ -85,12 +85,10 @@ public class DepartamentoController {
         File file = null;
 
         try {
-            String url;
-            if(   logo != null  && !logo.isEmpty()){ // si hay un logo, entonces hace esto
+            String url = null;
+            if( logo != null && !logo.isEmpty() ){ // si hay un logo, entonces hace esto
                 //return new ResponseEntity<>("Logo vacio", HttpStatus.UNPROCESSABLE_ENTITY);
                 url = logoServicio.save(logo);
-
-
                 // Encontrar el índice de "logo/"
                 int i = url.indexOf("logo/");
                 // Extraer la subcadena desde el índice i + 5 (para saltar el "logo/")
@@ -99,7 +97,8 @@ public class DepartamentoController {
                 file = new File("./uploads/" + sub);
 
             }else {
-                url = null;
+                    // Si no existe, asignar la url por defecto
+                    url = "https://upload.wikimedia.org/wikipedia/commons/thumb/3/3f/Placeholder_view_vector.svg/310px-Placeholder_view_vector.svg.png";
             }
 
 
@@ -123,11 +122,15 @@ public class DepartamentoController {
             return new ResponseEntity<>(mensaje, HttpStatus.UNPROCESSABLE_ENTITY);
         }
         catch (NullPointerException e) {
-            boolean borrado = file.delete();
+            if (file != null) {
+                boolean borrado = file.delete();
+            }
             return new ResponseEntity<>(e.getMessage() ,HttpStatus.BAD_REQUEST);
         }
         catch (Exception e) {
-            boolean borrado = file.delete();
+            if (file != null) {
+                boolean borrado = file.delete();
+            }
             return new ResponseEntity<>(e.getMessage() ,HttpStatus.BAD_REQUEST);
         }
